requirements.txt
ccxt==4.3.77
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
python-dotenv==1.0.1
tenacity==8.2.3

fastapi==0.112.2
uvicorn==0.30.5
jinja2==3.1.4
python-multipart==0.0.9

matplotlib==3.9.0

reportlab==3.6.13
